Refactor: Major UI architecture overhaul using MixIn pattern
Complete refactoring of AutoQAC_Interface.py from 961 lines to 84 lines using MixIn classes for better separation of concerns. Added new helper modules for game detection, plugin validation, process utilities, and configuration dialogs.

Main changes:
- Extract UI components into modular MixIn classes in AutoQACLib/ui/mixins/
- Move dialogs to dedicated files in AutoQACLib/ui/dialogs/
- Create MainWindow class using composition of focused mixins
- Add helper modules: game_detection, plugin_validator, process_utils, yaml_manager
- Extract configuration dialogs to separate module
- Update CLAUDE.md with file size limits and GUI development guidelines
- Fix documentation: xEdit uses -QAC flag, not -quickautoclean
- Add comprehensive refactoring plan documentation

File structure improvements:
- One class per file (550 line hard limit)
- Each mixin handles single responsibility
- Better testability and maintainability
- Cleaner dependency injection pattern

Coverage and test updates:
- Update test files for new architecture
- Maintain test coverage requirements
- Update coverage.xml with latest results
