+++ AutoQAC_Interface.py
@@ -1,36 +1,19 @@
 #!/usr/bin/env python3
-"""AutoQAC Main Interface - Refactored with centralized state management."""
+"""AutoQAC Main Interface - Slim entry point with refactored components."""
 
 from __future__ import annotations
 
 import sys
-from contextlib import suppress
 from pathlib import Path
 from typing import TYPE_CHECKING
 
-from PySide6.QtCore import QCoreApplication, Qt, QTimer, Slot
-from PySide6.QtGui import QAction, QCloseEvent, QFont
-from PySide6.QtWidgets import (
-    QApplication,
-    QDialog,
-    QDialogButtonBox,
-    QGroupBox,
-    QHBoxLayout,
-    QLabel,
-    QMainWindow,
-    QMenu,
-    QMenuBar,
-    QMessageBox,
-    QProgressBar,
-    QPushButton,
-    QVBoxLayout,
-    QWidget,
-)
+from PySide6.QtWidgets import QApplication
 
 from AutoQACLib.config_manager import ConfigManager
 from AutoQACLib.gui_controller import GuiController
 from AutoQACLib.logging_config import get_logger, log_startup_info, setup_logging
-from AutoQACLib.state_manager import AppState, StateManager
+from AutoQACLib.state_manager import StateManager
+from AutoQACLib.ui.main_window import MainWindow
 
 if TYPE_CHECKING:
     from logging import Logger
@@ -44,806 +27,6 @@ AUTOQAC_CONFIG_PATH: Path = AUTOQAC_DATA_PATH / "AutoQAC Config.yaml"  # New con
 setup_logging()
 logger: Logger = get_logger(__name__)
 log_startup_info(logger)
-
-
-class CleaningProgressDialog(QDialog):
-    """Dialog that shows cleaning progress and statistics."""
-
-    def __init__(self, parent: QWidget | None = None) -> None:
-        """Initialize the cleaning progress dialog."""
-        super().__init__(parent)
-        self._current_plugin_name: str | None = None
-        self.setWindowTitle("Cleaning Progress")
-        self.setMinimumSize(450, 300)
-        self.setModal(False)  # Non-modal so user can interact with main window
-
-        # Track if cleaning is in progress
-        self._cleaning_in_progress: bool = True
-
-        # Create UI elements
-        self._setup_ui()
-
-    def _setup_ui(self) -> None:
-        """Setup the dialog UI."""
-        layout: QVBoxLayout = QVBoxLayout(self)
-
-        # Progress section
-        progress_group: QGroupBox = QGroupBox("Progress")
-        progress_layout: QVBoxLayout = QVBoxLayout()
-
-        # Current plugin label
-        self.current_plugin_label: QLabel = QLabel("Waiting to start...")
-        self.current_plugin_label.setAlignment(Qt.AlignmentFlag.AlignCenter)
-        font: QFont = self.current_plugin_label.font()
-        font.setPointSize(font.pointSize() + 2)
-        self.current_plugin_label.setFont(font)
-        progress_layout.addWidget(self.current_plugin_label)
-
-        # Progress bar
-        self.progress_bar: QProgressBar = QProgressBar()
-        self.progress_bar.setTextVisible(True)
-        self.progress_bar.setFormat("%p%")
-        progress_layout.addWidget(self.progress_bar)
-
-        # Progress text
-        self.progress_label: QLabel = QLabel("0 / 0 plugins")
-        self.progress_label.setAlignment(Qt.AlignmentFlag.AlignCenter)
-        progress_layout.addWidget(self.progress_label)
-
-        progress_group.setLayout(progress_layout)
-        layout.addWidget(progress_group)
-
-        # Statistics section
-        stats_group: QGroupBox = QGroupBox("Statistics")
-        stats_layout: QVBoxLayout = QVBoxLayout()
-
-        # Create statistics labels
-        self.stats_labels: dict[str, QLabel] = {}
-        stats_items: list[tuple[str, str]] = [
-            ("cleaned", "✓ Cleaned:"),
-            ("failed", "✗ Failed:"),
-            ("skipped", "⊘ Skipped:"),
-            ("total", "Total:"),
-        ]
-
-        for key, label_text in stats_items:
-            row_layout: QHBoxLayout = QHBoxLayout()
-            label: QLabel = QLabel(label_text)
-            label.setMinimumWidth(150)
-            row_layout.addWidget(label)
-
-            value_label: QLabel = QLabel("0")
-            value_label.setAlignment(Qt.AlignmentFlag.AlignRight)
-            self.stats_labels[key] = value_label
-            row_layout.addWidget(value_label)
-            row_layout.addStretch()
-
-            stats_layout.addLayout(row_layout)
-
-        stats_group.setLayout(stats_layout)
-        layout.addWidget(stats_group)
-
-        layout.addStretch()
-
-        # Button box
-        self.button_box: QDialogButtonBox = QDialogButtonBox()
-
-        # Stop button (only shown during cleaning)
-        self.stop_button: QPushButton = QPushButton("Stop Cleaning")
-        self.stop_button.setStyleSheet("QPushButton { background-color: #ff4444; color: white; }")
-        self.button_box.addButton(self.stop_button, QDialogButtonBox.ButtonRole.ActionRole)
-
-        # Close button (only enabled after cleaning)
-        self.close_button: QPushButton = self.button_box.addButton(QDialogButtonBox.StandardButton.Close)
-        self.close_button.setEnabled(False)
-
-        self.button_box.rejected.connect(self.reject)
-        layout.addWidget(self.button_box)
-
-    def update_progress(self, current: int, total: int) -> None:
-        """
-        Updates the progress of a progress bar and its associated label based on the current and total values.
-
-        This method calculates the progress percentage, updates the progress bar's value, and sets the text of
-        the progress label to indicate the current and total progress. If a `_current_plugin_name` attribute
-        is present and non-empty, it incorporates the plugin name into the progress bar's displayed format.
-        If the total is zero, it resets the progress bar and label to their initial states.
-
-        Args:
-            current (int): The current progress count of the operation.
-            total (int): The total progress count for the operation.
-        """
-        if total > 0:
-            percentage: int = int((current / total) * 100)
-            self.progress_bar.setValue(percentage)
-            self.progress_label.setText(f"{current} / {total} plugins")
-
-            # Update progress bar format to show current plugin and percentage
-            if hasattr(self, "_current_plugin_name") and self._current_plugin_name:
-                self.progress_bar.setFormat(f"{self._current_plugin_name} / {percentage}%")
-            else:
-                self.progress_bar.setFormat("%p%")
-        else:
-            self.progress_bar.setValue(0)
-            self.progress_label.setText("0 / 0 plugins")
-            self.progress_bar.setFormat("%p%")
-
-    def update_current_plugin(self, plugin_name: str) -> None:
-        """
-        Updates the current plugin label and manages the progress bar display.
-
-        This function sets the label text to reflect the name of the currently processing
-        plugin. It additionally formats and updates the progress bar display if a progress
-        value greater than zero is present.
-
-        Args:
-            plugin_name: The name of the plugin to be set as the current processing plugin.
-        """
-        self.current_plugin_label.setText(f"Processing: {plugin_name}")
-        self._current_plugin_name = plugin_name
-
-        # Update progress bar format
-        if self.progress_bar.value() > 0:
-            self.progress_bar.setFormat(f"{plugin_name} / {self.progress_bar.value()}%")
-
-    def update_statistics(self, stats: dict[str, int]) -> None:
-        """
-        Updates the statistics display for the user interface.
-
-        This method iterates through the provided statistics dictionary and updates the
-        associated labels in the user interface with the corresponding values. The keys
-        in the given `stats` dictionary must match the keys defined in `self.stats_labels`
-        for the display to be updated.
-
-        Args:
-            stats: A dictionary where the keys are strings representing statistical
-                categories, and the values are integers representing the statistics
-                to be displayed.
-        """
-        for key, label in self.stats_labels.items():
-            if key in stats:
-                label.setText(str(stats[key]))
-
-    def set_cleaning_finished(self) -> None:
-        """
-        Marks the completion of the cleaning process and updates the UI components accordingly.
-
-        This method ensures the cleaning process is visually finalized by changing the UI elements.
-        It updates the text displayed on a label, hides the stop button, enables the close button,
-        and adjusts the progress bar format depending on the state of completion.
-
-        Returns:
-            None
-        """
-        self._cleaning_in_progress = False
-        self.current_plugin_label.setText("Cleaning completed!")
-        self.stop_button.setVisible(False)
-        self.close_button.setEnabled(True)
-
-        # Update progress bar format to show completion
-        if self.progress_bar.value() == 100:
-            self.progress_bar.setFormat("Completed - 100%")
-        else:
-            self.progress_bar.setFormat(f"Stopped - {self.progress_bar.value()}%")
-
-    def cleanup(self) -> None:
-        """Clean up signal connections."""
-        with suppress(RuntimeError):
-            self.button_box.rejected.disconnect()
-    def closeEvent(self, event: QCloseEvent) -> None:
-        """Handle close event."""
-        if self._cleaning_in_progress:
-            reply: QMessageBox.StandardButton = QMessageBox.question(
-                self,
-                "Cleaning in Progress",
-                "Cleaning is still in progress. Are you sure you want to close this dialog?\n\n"
-                "Note: Closing this dialog will not stop the cleaning process.",
-                QMessageBox.StandardButton.Yes | QMessageBox.StandardButton.No,
-                QMessageBox.StandardButton.No,
-            )
-            if reply == QMessageBox.StandardButton.No:
-                event.ignore()
-                return
-        self.cleanup()
-        event.accept()
-
-
-class MainWindow(QMainWindow):
-    """Main application window with refactored state management."""
-
-    def __init__(self, state: StateManager, controller: GuiController) -> None:
-        """Initialize the main window."""
-        super().__init__()
-        self.state: StateManager = state
-        self.controller: GuiController = controller
-        self._updating_ui: bool = False  # Flag to prevent recursive updates
-        self._update_timer: QTimer = QTimer(self)  # Timer for debouncing UI updates
-        self._update_timer.setSingleShot(True)
-        self._update_timer.setInterval(50)  # 50ms delay
-        self._update_timer.timeout.connect(self._perform_ui_update)
-        self._update_priority: str = "normal"  # Track update priority
-
-        # UI elements
-        self.progress_dialog: CleaningProgressDialog | None = None
-        self.load_order_button: QPushButton | None = None
-        self.mo2_button: QPushButton | None = None
-        self.mo2_mode_button: QPushButton | None = None
-        self.partial_forms_button: QPushButton | None = None
-        self.xedit_button: QPushButton | None = None
-        self.start_button: QPushButton | None = None
-        self.stop_button: QPushButton | None = None
-
-        # Connect controller signals
-        self._connect_controller_signals()
-
-        # Connect state signals
-        self._connect_state_signals()
-
-        # Setup UI
-        self._setup_ui()
-
-        # Update initial state
-        self._update_ui_from_state()
-
-    def _connect_controller_signals(self) -> None:
-        """Connect controller signals to UI updates."""
-        self.controller.show_message.connect(self._show_message)
-        self.controller.show_error.connect(self._show_error)
-        self.controller.update_status.connect(self._update_status)
-
-    def _connect_state_signals(self) -> None:
-        """Connect state signals to UI updates."""
-        self.state.configuration_changed.connect(self._on_configuration_changed)
-        self.state.progress_changed.connect(self._on_progress_changed)
-        self.state.cleaning_started.connect(self._on_cleaning_started)
-        self.state.cleaning_finished.connect(self._on_cleaning_finished)
-        self.state.plugin_processed.connect(self._on_plugin_processed)
-        self.state.state_changed.connect(self._on_state_changed)
-        self.state.bulk_state_changed.connect(self._on_bulk_state_changed)
-
-    def _setup_ui(self) -> None:
-        """Setup the main window UI."""
-        self.setWindowTitle("AutoQAC - Refactored")
-        self.setMinimumSize(600, 400)
-
-        # Create central widget
-        central_widget: QWidget = QWidget(self)
-        self.setCentralWidget(central_widget)
-
-        # Create main layout
-        main_layout: QVBoxLayout = QVBoxLayout(central_widget)
-
-        # Create configuration group
-        config_group: QGroupBox = self._create_configuration_group()
-        main_layout.addWidget(config_group)
-
-        # Create control group
-        control_group: QGroupBox = self._create_control_group()
-        main_layout.addWidget(control_group)
-
-        # Create menu bar
-        self._create_menu_bar()
-
-        # Create status bar
-        self.status_bar = self.statusBar()
-        self.status_bar.showMessage("Ready")
-
-        # Initial UI update
-        self._update_ui_from_state()
-
-    def _create_configuration_group(self) -> QGroupBox:
-        """Create the configuration section."""
-        group: QGroupBox = QGroupBox("Configuration")
-        layout: QVBoxLayout = QVBoxLayout()
-
-        # Load Order
-        lo_layout: QHBoxLayout = QHBoxLayout()
-        self.load_order_button = QPushButton("Configure Load Order")
-        self.load_order_button.clicked.connect(self._configure_load_order)
-        lo_layout.addWidget(self.load_order_button)
-        lo_layout.addStretch()
-        layout.addLayout(lo_layout)
-
-        # MO2
-        mo2_layout: QHBoxLayout = QHBoxLayout()
-        self.mo2_button = QPushButton("Configure MO2")
-        self.mo2_button.clicked.connect(self._configure_mo2)
-        mo2_layout.addWidget(self.mo2_button)
-
-        self.mo2_mode_button = QPushButton("MO2 Mode: OFF")
-        self.mo2_mode_button.setCheckable(True)
-        self.mo2_mode_button.clicked.connect(self._toggle_mo2_mode)
-        mo2_layout.addWidget(self.mo2_mode_button)
-        mo2_layout.addStretch()
-        layout.addLayout(mo2_layout)
-
-        # xEdit
-        xedit_layout: QHBoxLayout = QHBoxLayout()
-        self.xedit_button = QPushButton("Configure xEdit")
-        self.xedit_button.clicked.connect(self._configure_xedit)
-        xedit_layout.addWidget(self.xedit_button)
-        xedit_layout.addStretch()
-        layout.addLayout(xedit_layout)
-
-        # Partial Forms (Experimental Feature)
-        partial_forms_layout: QHBoxLayout = QHBoxLayout()
-        self.partial_forms_button = QPushButton("Partial Forms: OFF")
-        self.partial_forms_button.setCheckable(True)
-        self.partial_forms_button.setToolTip("Enable experimental Partial Forms feature (requires xEdit >= 4.1.5b)")
-        self.partial_forms_button.clicked.connect(self._toggle_partial_forms)
-        partial_forms_layout.addWidget(self.partial_forms_button)
-        partial_forms_layout.addStretch()
-        layout.addLayout(partial_forms_layout)
-
-        group.setLayout(layout)
-        return group
-
-    def _create_control_group(self) -> QGroupBox:
-        """Create the control section."""
-        group: QGroupBox = QGroupBox("Controls")
-        layout: QHBoxLayout = QHBoxLayout()
-
-        self.start_button = QPushButton("Start Cleaning")
-        self.start_button.clicked.connect(self._start_cleaning)
-        layout.addWidget(self.start_button)
-
-        self.stop_button = QPushButton("Stop")
-        self.stop_button.clicked.connect(self._stop_cleaning)
-        self.stop_button.setEnabled(False)
-        layout.addWidget(self.stop_button)
-
-        layout.addStretch()
-
-        group.setLayout(layout)
-        return group
-
-    def _create_menu_bar(self) -> None:
-        """Create the menu bar."""
-        menubar: QMenuBar = self.menuBar()
-
-        # File menu
-        file_menu: QMenu = menubar.addMenu("&File")
-
-        refresh_action: QAction = QAction("&Refresh Configuration", self)
-        refresh_action.triggered.connect(self.controller.refresh_configuration)
-        file_menu.addAction(refresh_action)
-
-        file_menu.addSeparator()
-
-        exit_action: QAction = QAction("E&xit", self)
-        exit_action.triggered.connect(self.close)
-        file_menu.addAction(exit_action)
-
-        # Help menu
-        help_menu: QMenu = menubar.addMenu("&Help")
-
-        about_action: QAction = QAction("&About", self)
-        about_action.triggered.connect(self._show_about)
-        help_menu.addAction(about_action)
-
-    def _update_ui_from_state(self, priority: str = "normal") -> None:
-        """Update UI elements based on current state with priority-based debouncing.
-        
-        Args:
-            priority: Update priority - "immediate", "high", or "normal"
-        """
-        # Immediate updates bypass debouncing
-        if priority == "immediate":
-            self._perform_ui_update()
-            return
-        
-        # Set appropriate debounce interval based on priority
-        interval = 50 if priority == "high" else 200  # Longer delay for normal updates
-        
-        # Update priority if this is a higher priority update
-        if priority == "high" and self._update_priority == "normal":
-            self._update_priority = priority
-            self._update_timer.setInterval(interval)
-        
-        # Start or restart timer with appropriate interval
-        if not self._update_timer.isActive() or priority == "high":
-            self._update_timer.setInterval(interval)
-            self._update_timer.start()
-
-    def _perform_ui_update(self) -> None:
-        """Perform the actual UI update with debouncing."""
-        if self._updating_ui:
-            return  # Prevent recursive updates
-
-        self._updating_ui = True
-        self._update_priority = "normal"  # Reset priority after update
-        try:
-            state_snapshot: AppState = self.state.state
-
-            # Update configuration buttons efficiently
-            if self.load_order_button:
-                new_text = "Load Order ✓" if state_snapshot.is_load_order_configured else "Configure Load Order"
-                if self.load_order_button.text() != new_text:
-                    self._update_button_state(
-                        self.load_order_button,
-                        state_snapshot.is_load_order_configured,
-                        new_text,
-                    )
-
-            if self.mo2_button:
-                new_text = "MO2 ✓" if state_snapshot.is_mo2_configured else "Configure MO2"
-                if self.mo2_button.text() != new_text:
-                    self._update_button_state(
-                        self.mo2_button,
-                        state_snapshot.is_mo2_configured,
-                        new_text,
-                    )
-
-            if self.xedit_button:
-                new_text = "xEdit ✓" if state_snapshot.is_xedit_configured else "Configure xEdit"
-                if self.xedit_button.text() != new_text:
-                    self._update_button_state(
-                        self.xedit_button,
-                        state_snapshot.is_xedit_configured,
-                        new_text,
-                    )
-
-            # Update MO2 mode button efficiently
-            if self.mo2_mode_button:
-                new_checked = state_snapshot.mo2_mode
-                if self.mo2_mode_button.isChecked() != new_checked:
-                    self.mo2_mode_button.setChecked(new_checked)
-
-                new_text = f"MO2 Mode: {'ON' if state_snapshot.mo2_mode else 'OFF'}"
-                if self.mo2_mode_button.text() != new_text:
-                    self.mo2_mode_button.setText(new_text)
-
-            # Update Partial Forms button efficiently
-            if self.partial_forms_button:
-                new_checked = state_snapshot.partial_forms_enabled
-                if self.partial_forms_button.isChecked() != new_checked:
-                    self.partial_forms_button.setChecked(new_checked)
-
-                new_text = f"Partial Forms: {'ON' if state_snapshot.partial_forms_enabled else 'OFF'}"
-                if self.partial_forms_button.text() != new_text:
-                    self.partial_forms_button.setText(new_text)
-
-            # Update control buttons efficiently
-            if self.start_button:
-                new_enabled = state_snapshot.is_fully_configured and not state_snapshot.is_cleaning
-                if self.start_button.isEnabled() != new_enabled:
-                    self.start_button.setEnabled(new_enabled)
-
-            if self.stop_button:
-                new_enabled = state_snapshot.is_cleaning
-                if self.stop_button.isEnabled() != new_enabled:
-                    self.stop_button.setEnabled(new_enabled)
-        finally:
-            self._updating_ui = False
-
-    @staticmethod
-    def _update_button_state(button: QPushButton, configured: bool, text: str) -> None:
-        """Update button appearance based on configuration state."""
-        button.setText(text)
-        if configured:
-            button.setStyleSheet("QPushButton { background-color: #4CAF50; color: white; }")
-        else:
-            button.setStyleSheet("")
-
-    @Slot()
-    def _configure_load_order(self) -> None:
-        """Configure load order file."""
-        self.controller.configure_load_order(self)
-
-    @Slot()
-    def _configure_mo2(self) -> None:
-        """Configure MO2."""
-        self.controller.configure_mo2(self)
-
-    @Slot()
-    def _configure_xedit(self) -> None:
-        """Configure xEdit."""
-        self.controller.configure_xedit(self)
-
-    @Slot()
-    def _toggle_mo2_mode(self) -> None:
-        """Toggle MO2 mode."""
-        if self.mo2_mode_button is None:
-            return
-        enabled: bool = self.mo2_mode_button.isChecked()
-        self.controller.toggle_mo2_mode(enabled)
-
-    @Slot()
-    def _toggle_partial_forms(self) -> None:
-        """Toggle partial forms cleaning."""
-        if self.partial_forms_button is None:
-            return
-        enabled: bool = self.partial_forms_button.isChecked()
-        if enabled:
-            # Show warning dialog if enabling
-            confirmed: bool = show_partial_forms_warning(self)
-            if not confirmed:
-                # User cancelled or closed dialog, revert button
-                self.partial_forms_button.setChecked(False)
-                return
-        self.controller.toggle_partial_forms(enabled)
-
-    @Slot()
-    def _start_cleaning(self) -> None:
-        """Start the cleaning process."""
-        try:
-            logger.info("Starting cleaning process...")
-            self.controller.start_cleaning()
-        except (OSError, RuntimeError, ValueError) as e:
-            logger.error(f"Error starting cleaning: {e}")
-            self._show_error("Error", f"Failed to start cleaning: {e}")
-
-    @Slot()
-    def _stop_cleaning(self) -> None:
-        """Stop the cleaning process."""
-        self.controller.stop_cleaning()
-
-    @Slot(bool)
-    def _on_configuration_changed(self, is_fully_configured: bool) -> None:
-        """Handle configuration changes."""
-        if self.start_button is None:
-            return
-        self.start_button.setEnabled(is_fully_configured and not self.state.get("is_cleaning"))
-        if is_fully_configured:
-            self._log("All paths configured ✓")
-
-    @Slot(int, int)
-    def _on_progress_changed(self, current: int, total: int) -> None:
-        """Handle progress updates."""
-        if total > 0:
-            percentage: float = (current / total) * 100
-            self._update_status(f"Progress: {current}/{total} ({percentage:.1f}%)")
-
-        # Update progress dialog
-        if self.progress_dialog and self.progress_dialog.isVisible():
-            self.progress_dialog.update_progress(current, total)
-            # Update statistics
-            stats: dict[str, int] = self.state.state.cleaning_stats
-            self.progress_dialog.update_statistics(stats)
-
-    @Slot()
-    def _on_cleaning_started(self) -> None:
-        """Handle cleaning start."""
-        if self.start_button is None or self.stop_button is None:
-            return
-
-        # Create and show progress dialog
-        self.progress_dialog = CleaningProgressDialog(self)
-        self.progress_dialog.stop_button.clicked.connect(self._stop_cleaning)
-
-        # Update progress with current state
-        stats: dict[str, int] = self.state.state.cleaning_stats
-        self.progress_dialog.update_statistics(stats)
-        self.progress_dialog.update_progress(self.state.get("progress", 0), self.state.get("total_plugins", 0))
-
-        self.progress_dialog.show()
-        self._log("Cleaning started...")
-        self.start_button.setEnabled(False)
-        self.stop_button.setEnabled(True)
-
-    @Slot()
-    def _on_cleaning_finished(self) -> None:
-        """Handle cleaning completion."""
-        if self.start_button is None or self.stop_button is None:
-            return
-
-        self._log("Cleaning finished!")
-
-        # Update progress dialog
-        if self.progress_dialog:
-            self.progress_dialog.set_cleaning_finished()
-            # Update final statistics
-            stats: dict[str, int] = self.state.state.cleaning_stats
-            self.progress_dialog.update_statistics(stats)
-
-        self.start_button.setEnabled(True)
-        self.stop_button.setEnabled(False)
-
-    @Slot(str, str, str)
-    def _on_plugin_processed(self, plugin: str, status: str, message: str) -> None:
-        """Handle plugin processing completion."""
-        icon: str = {
-            "cleaned": "✓",
-            "failed": "✗",
-            "skipped": "⊘",
-        }.get(status, "?")
-
-        self._log(f"{icon} {plugin}: {message}")
-
-        # Update current plugin in progress dialog
-        if self.progress_dialog and self.progress_dialog.isVisible():
-            current_plugin: str | None = self.state.get("current_plugin")
-            if current_plugin:
-                self.progress_dialog.update_current_plugin(current_plugin)
-
-    @Slot(dict)
-    def _on_bulk_state_changed(self, changes: dict[str, object]) -> None:
-        """Handle multiple state changes at once for better performance."""
-        # Check if any changes require immediate update
-        immediate_update = False
-        high_priority_update = False
-        
-        for property_name, value in changes.items():
-            if property_name == "is_cleaning":
-                immediate_update = True
-                if value:
-                    self._on_cleaning_started()
-                else:
-                    self._on_cleaning_finished()
-            elif property_name in ["current_plugin", "current_operation"]:
-                high_priority_update = True
-                if (
-                    property_name == "current_plugin"
-                    and self.progress_dialog
-                    and self.progress_dialog.isVisible()
-                    and isinstance(value, str)
-                ):
-                    self.progress_dialog.update_current_plugin(value)
-        
-        # Update UI with appropriate priority
-        if immediate_update:
-            self._update_ui_from_state("immediate")
-        elif high_priority_update:
-            self._update_ui_from_state("high")
-        else:
-            self._update_ui_from_state("normal")
-    
-    @Slot(str, object)
-    def _on_state_changed(self, property_name: str, value: object) -> None:
-        """Handle individual state property changes with appropriate priority."""
-        # Determine update priority based on property type
-        priority = "normal"
-        
-        # High priority for user-facing state changes
-        if property_name == "is_cleaning":
-            priority = "immediate"  # Cleaning state changes need immediate feedback
-        elif property_name in ["current_plugin", "current_operation"]:
-            priority = "high"  # Plugin progress updates
-            # Update current plugin in progress dialog immediately
-            if (
-                property_name == "current_plugin"
-                and self.progress_dialog
-                and self.progress_dialog.isVisible()
-                and isinstance(value, str)
-            ):
-                self.progress_dialog.update_current_plugin(value)
-        elif property_name in ["mo2_mode", "partial_forms_enabled"]:
-            priority = "normal"  # Settings changes
-        
-        # Update UI with appropriate priority
-        if property_name in [
-            "is_load_order_configured",
-            "is_mo2_configured",
-            "is_xedit_configured",
-            "mo2_mode",
-            "partial_forms_enabled",
-            "is_cleaning",
-        ]:
-            self._update_ui_from_state(priority)
-
-    @Slot(str, str)
-    def _show_message(self, title: str, message: str) -> None:
-        """Show an information message."""
-        QMessageBox.information(self, title, message)
-
-    @Slot(str, str)
-    def _show_error(self, title: str, message: str) -> None:
-        """Show an error message."""
-        QMessageBox.critical(self, title, message)
-
-    @Slot(str)
-    def _update_status(self, message: str) -> None:
-        """Update the status bar."""
-        if self.status_bar is None:
-            return
-        self.status_bar.showMessage(message)
-
-    def _log(self, message: str) -> None:
-        """Add a message to the log display."""
-        # Log messages are now only shown in status bar during cleaning
-        if self.state.get("is_cleaning"):
-            self._update_status(message)
-
-    def _show_about(self) -> None:
-        """Show the about dialog."""
-        QMessageBox.about(
-            self,
-            "About AutoQAC",
-            "AutoQAC - Automated Quick Auto Clean\n\n"
-            "A PySide6 application for batch cleaning Bethesda game plugins "
-            "using xEdit's quickautoclean functionality.\n\n"
-            "Version: 2.0.0\n"
-            "Built with PySide6 and Python 3.8+",
-        )
-
-    def _disconnect_all_signals(self) -> None:
-        """Disconnect all signal connections to prevent memory leaks."""
-        # Disconnect controller signals
-        try:
-            self.controller.show_message.disconnect(self._show_message)
-            self.controller.show_error.disconnect(self._show_error)
-            self.controller.update_status.disconnect(self._update_status)
-        except RuntimeError:
-            pass  # Already disconnected
-
-        # Disconnect state signals
-        try:
-            self.state.configuration_changed.disconnect(self._on_configuration_changed)
-            self.state.progress_changed.disconnect(self._on_progress_changed)
-            self.state.cleaning_started.disconnect(self._on_cleaning_started)
-            self.state.cleaning_finished.disconnect(self._on_cleaning_finished)
-            self.state.plugin_processed.disconnect(self._on_plugin_processed)
-            self.state.state_changed.disconnect(self._on_state_changed)
-        except RuntimeError:
-            pass  # Already disconnected
-
-        # Disconnect button signals
-        if self.load_order_button:
-            with suppress(RuntimeError):
-                self.load_order_button.clicked.disconnect()
-        if self.mo2_button:
-            with suppress(RuntimeError):
-                self.mo2_button.clicked.disconnect()
-        if self.mo2_mode_button:
-            with suppress(RuntimeError):
-                self.mo2_mode_button.clicked.disconnect()
-        if self.xedit_button:
-            with suppress(RuntimeError):
-                self.xedit_button.clicked.disconnect()
-        if self.partial_forms_button:
-            with suppress(RuntimeError):
-                self.partial_forms_button.clicked.disconnect()
-        if self.start_button:
-            with suppress(RuntimeError):
-                self.start_button.clicked.disconnect()
-        if self.stop_button:
-            with suppress(RuntimeError):
-                self.stop_button.clicked.disconnect()
-
-        # Disconnect timer
-        with suppress(RuntimeError):
-            self._update_timer.timeout.disconnect()
-
-    def closeEvent(self, event: QCloseEvent) -> None:
-        """Handle window close event with proper cleanup."""
-        # If cleaning is in progress, ask user for confirmation
-        if self.state.get("is_cleaning"):
-            reply: QMessageBox.StandardButton = QMessageBox.question(
-                self,
-                "Cleaning in Progress",
-                "Cleaning is currently in progress. Are you sure you want to close the application?\n\n"
-                "This will stop the cleaning process.",
-                QMessageBox.StandardButton.Yes | QMessageBox.StandardButton.No,
-                QMessageBox.StandardButton.No,
-            )
-            if reply == QMessageBox.StandardButton.No:
-                event.ignore()
-                return
-
-        # Stop update timer
-        if self._update_timer.isActive():
-            self._update_timer.stop()
-
-        # Disconnect all signals
-        self._disconnect_all_signals()
-
-        # Close progress dialog if open
-        if self.progress_dialog:
-            if self.progress_dialog.isVisible():
-                self.progress_dialog.close()
-            self.progress_dialog.cleanup()
-            self.progress_dialog.deleteLater()
-            self.progress_dialog = None
-
-        # Perform comprehensive cleanup
-        logger.info("Shutting down application")
-        self.controller.cleanup()
-
-        # Ensure all events are processed
-        QCoreApplication.processEvents()
-
-        event.accept()
-        logger.info("Application shutdown complete")
 
 
 def create_application() -> tuple[QApplication, MainWindow]:
@@ -899,62 +82,3 @@ def main() -> None:
 
 if __name__ == "__main__":
     main()
-
-
-def show_partial_forms_warning(parent: QWidget) -> bool:
-    """
-    Show a warning dialog about partial forms cleaning.
-
-    Args:
-        parent: The parent widget for the dialog.
-
-    Returns:
-        True if the user wants to enable partial forms cleaning, False otherwise.
-    """
-    dialog: QDialog = QDialog(parent)
-    dialog.setWindowTitle("Partial Forms Warning")
-    dialog.setModal(True)
-    dialog.setMinimumSize(400, 200)
-
-    layout: QVBoxLayout = QVBoxLayout(dialog)
-
-    # Warning message
-    warning_label: QLabel = QLabel(
-        "⚠️  WARNING: Partial Forms Cleaning\n\n"
-        "Enabling partial forms cleaning will clean plugins that contain "
-        "partial forms (records that are split across multiple plugins).\n\n"
-        "This can be dangerous and may break your game if not done carefully.\n\n"
-        "Only enable this if you understand the risks and have a backup "
-        "of your plugins.\n\n"
-        "Do you want to enable partial forms cleaning?"
-    )
-    warning_label.setWordWrap(True)
-    layout.addWidget(warning_label)
-
-    # Button box
-    button_box: QDialogButtonBox = QDialogButtonBox()
-    enable_button: QPushButton = QPushButton("Enable Partial Forms")
-    enable_button.setStyleSheet("QPushButton { background-color: #ff4444; color: white; }")
-    cancel_button: QPushButton = QPushButton("Cancel")
-    cancel_button.setStyleSheet("QPushButton { background-color: #666666; color: white; }")
-
-    button_box.addButton(enable_button, QDialogButtonBox.ButtonRole.AcceptRole)
-    button_box.addButton(cancel_button, QDialogButtonBox.ButtonRole.RejectRole)
-
-    layout.addWidget(button_box)
-
-    # Connect signals
-    def on_enable() -> None:
-        dialog.accept()
-
-    def on_cancel() -> None:
-        dialog.reject()
-
-    enable_button.clicked.connect(on_enable)
-    cancel_button.clicked.connect(on_cancel)
-    button_box.accepted.connect(on_enable)
-    button_box.rejected.connect(on_cancel)
-
-    # Show dialog and return result
-    result: int = dialog.exec()
-    return result == QDialog.DialogCode.Accepted
